address review: merge gossip harness module, tidy features - Combine gossip harness + conversion into a single conformance/src/gossip.rs - ffi now requires agave-unstable-api; drop unused frozen-abi feature - Drop dummy-for-ci-check from solana-gossip (covered by CI dcou runs) - Make qualifier_attr an optional dep enabled by dev-context-only-utils
