+++ gossip/Cargo.toml
@@ -37,9 +37,9 @@ frozen-abi = [
 agave-unstable-api = []
 dev-context-only-utils = [
     "agave-votor-messages/dev-context-only-utils",
+    "dep:qualifier_attr",
     "solana-bloom/dev-context-only-utils",
 ]
-dummy-for-ci-check = ["dev-context-only-utils"]
 
 [dependencies]
 agave-logger = { workspace = true }
@@ -57,7 +57,7 @@ itertools = { workspace = true }
 lazy-lru = { workspace = true }
 log = { workspace = true }
 num-traits = { workspace = true }
-qualifier_attr = { workspace = true }
+qualifier_attr = { workspace = true, optional = true }
 rand = { workspace = true }
 rand_chacha = { workspace = true }
 rayon = { workspace = true }
